moving profiling to builder. This makes more sense now, however logging and such could go to its own lower level module if need be.
